Add remoteName filed to Git struct
Added a new field to the Git struct: remoteName. With this we have the branch and remote configured at the start of the main function. This also permits us to not pass these values to the Push methods. I think it's cleaner and avoids any future mistakes.
Push method now build the commandArgs it needs to pass to the executeCommand function. It's now using the branch and remoteName fields in the Git struct.

Signed-off-by: Froilan Irizarry <froilan@froilanirizarry.me>